requirements.txt
alembic==1.16.4
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
blinker==1.9.0
click==8.2.1
colorama==0.4.6
fastapi==0.116.1
greenlet==3.2.3
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
python-multipart==0.0.20
qrcode==8.2
sniffio==1.3.1
SQLAlchemy==2.0.42
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
Werkzeug==3.1.3
sentence-transformers
faiss-cpu
numpy
langdetect
google-cloud-speech
google-cloud-texttospeech
opencv-python
imageio
lpips
boto3
aiofiles
google-generativeai
google-genai